+++ app/src/main/java/com/endorphinapps/kemikal/queenofclean/EditRecords/EditJob.java
@@ -653,11 +653,11 @@ public class EditJob extends MenuMain
      *
      * @return estimatedTime as an int
      */
-    private int getEstimatedTime() {
-        int estimatedTime = 0;
+    private double getEstimatedTime() {
+        double estimatedTime = 0;
 
         if (!et_estimatedTime.getText().toString().equals("")) {
-            estimatedTime = Integer.valueOf(et_estimatedTime.getText().toString());
+            estimatedTime = Double.valueOf(et_estimatedTime.getText().toString());
         }
         System.out.println("z! AddJob - getEstimatedTime(): " +
                 String.valueOf(estimatedTime));
@@ -721,7 +721,7 @@ public class EditJob extends MenuMain
         double priceOfJobItems = calculatePriceOfJobItems();
 
         //Get estimated employee time required
-        int estimateJobTime = getEstimatedTime();
+        double estimateJobTime = getEstimatedTime();
 
         //Calculate pay to employee
         double payToEmployee = calculatePayToEmployee();
@@ -790,7 +790,7 @@ public class EditJob extends MenuMain
      */
     private void saveToDB(long id, long startDate,
                           long startTime, String jobStatus, String customerPaymentStatus,
-                          String employeePaymentStatus, int estimateJobTime,
+                          String employeePaymentStatus, double estimateJobTime,
                           double totalCostForJob, String jobNotes,
                           long customerId, long employeeId) {
         db.updateJob(id, startDate, startTime,
